+++ CMakeLists.txt
@@ -246,7 +246,6 @@ rocm_package_setup_component(tests
 add_subdirectory(library)
 add_subdirectory(example)
 add_subdirectory(test)
-add_subdirectory(profiler)
 
 #Create an interface target for the include only files and call it "composablekernels"
 include(CMakePackageConfigHelpers)
