+++ library/CMakeLists.txt
@@ -1,3 +1,4 @@
-add_subdirectory(src/host_tensor)
 add_subdirectory(src/tensor_operation_instance/gpu)
+add_subdirectory(src/host_tensor)
 add_subdirectory(src/utility)
+add_subdirectory(src/profiler)
