+++ library/src/profiler/profile_batched_gemm.cpp
@@ -7,7 +7,7 @@
 #include <initializer_list>
 #include <cstdlib>
 
-#include "profiler/include/profile_batched_gemm_impl.hpp"
+#include "profile_batched_gemm_impl.hpp"
 
 enum struct GemmMatrixLayout
 {
