+++ library/src/profiler/profile_batched_gemm_reduce.cpp
@@ -6,9 +6,9 @@
 #include <initializer_list>
 #include <cstdlib>
 
-#include "profiler/include/profile_gemm_reduce_impl.hpp"
+#include "profile_batched_gemm_reduce_impl.hpp"
 
-int profile_gemm_reduce(int argc, char* argv[])
+int profile_batched_gemm_reduce(int argc, char* argv[])
 {
     enum struct GemmMatrixLayout
     {
@@ -24,9 +24,9 @@ int profile_gemm_reduce(int argc, char* argv[])
         F16_F16_F16_F32_F32, // 1
     };
 
-    if(!(argc == 14 || argc == 15))
+    if(!(argc == 15 || argc == 16))
     {
-        printf("arg1: tensor operation (gemm: GEMM+Reduce)\n");
+        printf("arg1: tensor operation (batched_gemm: BatchedGEMM+Reduce)\n");
         printf("arg2: data type (0: fp32; 1: fp16)\n");
         printf("arg3: matrix layout (0: A[m, k] * B[k, n] = C[m, n];\n");
         printf("                     1: A[m, k] * B[n, k] = C[m, n];\n");
@@ -36,8 +36,8 @@ int profile_gemm_reduce(int argc, char* argv[])
         printf("arg5: initialization (0: no init; 1: integer value; 2: decimal value)\n");
         printf("arg6: print tensor value (0: no; 1: yes)\n");
         printf("arg7: time kernel (0=n0, 1=yes)\n");
-        printf("arg8 to 13: M, N, K, StrideA, StrideB, StrideC\n");
-        printf("arg14: split k into  mulitiple batch\n");
+        printf("arg8 to 14: M, N, K, StrideA, StrideB, StrideC, BatchCount\n");
+        printf("arg15: split k into  mulitiple batch\n");
         exit(1);
     }
 
@@ -56,15 +56,17 @@ int profile_gemm_reduce(int argc, char* argv[])
     const int StrideB = std::stoi(argv[12]);
     const int StrideC = std::stoi(argv[13]);
 
+    const int BatchCount = std::stoi(argv[14]);
+
     if(data_type == GemmReduceDataType::F16_F16_F16_F32_F32 && layout == GemmMatrixLayout::MK_KN_MN)
     {
-        ck::profiler::profile_gemm_reduce_impl<ck::half_t,
-                                               ck::half_t,
-                                               ck::half_t,
-                                               float,
-                                               ck::tensor_layout::gemm::RowMajor,
-                                               ck::tensor_layout::gemm::RowMajor,
-                                               ck::tensor_layout::gemm::RowMajor>(
+        ck::profiler::profile_batched_gemm_reduce_impl<ck::half_t,
+                                                       ck::half_t,
+                                                       ck::half_t,
+                                                       float,
+                                                       ck::tensor_layout::gemm::RowMajor,
+                                                       ck::tensor_layout::gemm::RowMajor,
+                                                       ck::tensor_layout::gemm::RowMajor>(
             do_verification,
             init_method,
             do_log,
@@ -74,18 +76,19 @@ int profile_gemm_reduce(int argc, char* argv[])
             K,
             (StrideA < 0) ? K : StrideA,
             (StrideB < 0) ? N : StrideB,
-            (StrideC < 0) ? N : StrideC);
+            (StrideC < 0) ? N : StrideC,
+            BatchCount);
     }
     else if(data_type == GemmReduceDataType::F16_F16_F16_F32_F32 &&
             layout == GemmMatrixLayout::MK_NK_MN)
     {
-        ck::profiler::profile_gemm_reduce_impl<ck::half_t,
-                                               ck::half_t,
-                                               ck::half_t,
-                                               float,
-                                               ck::tensor_layout::gemm::RowMajor,
-                                               ck::tensor_layout::gemm::ColumnMajor,
-                                               ck::tensor_layout::gemm::RowMajor>(
+        ck::profiler::profile_batched_gemm_reduce_impl<ck::half_t,
+                                                       ck::half_t,
+                                                       ck::half_t,
+                                                       float,
+                                                       ck::tensor_layout::gemm::RowMajor,
+                                                       ck::tensor_layout::gemm::ColumnMajor,
+                                                       ck::tensor_layout::gemm::RowMajor>(
             do_verification,
             init_method,
             do_log,
@@ -95,18 +98,19 @@ int profile_gemm_reduce(int argc, char* argv[])
             K,
             (StrideA < 0) ? K : StrideA,
             (StrideB < 0) ? K : StrideB,
-            (StrideC < 0) ? N : StrideC);
+            (StrideC < 0) ? N : StrideC,
+            BatchCount);
     }
     else if(data_type == GemmReduceDataType::F16_F16_F16_F32_F32 &&
             layout == GemmMatrixLayout::KM_KN_MN)
     {
-        ck::profiler::profile_gemm_reduce_impl<ck::half_t,
-                                               ck::half_t,
-                                               ck::half_t,
-                                               float,
-                                               ck::tensor_layout::gemm::ColumnMajor,
-                                               ck::tensor_layout::gemm::RowMajor,
-                                               ck::tensor_layout::gemm::RowMajor>(
+        ck::profiler::profile_batched_gemm_reduce_impl<ck::half_t,
+                                                       ck::half_t,
+                                                       ck::half_t,
+                                                       float,
+                                                       ck::tensor_layout::gemm::ColumnMajor,
+                                                       ck::tensor_layout::gemm::RowMajor,
+                                                       ck::tensor_layout::gemm::RowMajor>(
             do_verification,
             init_method,
             do_log,
@@ -116,18 +120,19 @@ int profile_gemm_reduce(int argc, char* argv[])
             K,
             (StrideA < 0) ? M : StrideA,
             (StrideB < 0) ? N : StrideB,
-            (StrideC < 0) ? N : StrideC);
+            (StrideC < 0) ? N : StrideC,
+            BatchCount);
     }
     else if(data_type == GemmReduceDataType::F16_F16_F16_F32_F32 &&
             layout == GemmMatrixLayout::KM_NK_MN)
     {
-        ck::profiler::profile_gemm_reduce_impl<ck::half_t,
-                                               ck::half_t,
-                                               ck::half_t,
-                                               float,
-                                               ck::tensor_layout::gemm::ColumnMajor,
-                                               ck::tensor_layout::gemm::ColumnMajor,
-                                               ck::tensor_layout::gemm::RowMajor>(
+        ck::profiler::profile_batched_gemm_reduce_impl<ck::half_t,
+                                                       ck::half_t,
+                                                       ck::half_t,
+                                                       float,
+                                                       ck::tensor_layout::gemm::ColumnMajor,
+                                                       ck::tensor_layout::gemm::ColumnMajor,
+                                                       ck::tensor_layout::gemm::RowMajor>(
             do_verification,
             init_method,
             do_log,
@@ -137,7 +142,8 @@ int profile_gemm_reduce(int argc, char* argv[])
             K,
             (StrideA < 0) ? M : StrideA,
             (StrideB < 0) ? K : StrideB,
-            (StrideC < 0) ? N : StrideC);
+            (StrideC < 0) ? N : StrideC,
+            BatchCount);
     }
     else
     {
