+++ library/src/profiler/profile_conv_fwd_bias_relu.cpp
@@ -6,7 +6,7 @@
 #include <initializer_list>
 #include <cstdlib>
 
-#include "profiler/include/profile_conv_fwd_bias_relu_add_impl.hpp"
+#include "profile_conv_fwd_bias_relu_impl.hpp"
 
 enum struct ConvDataType
 {
@@ -32,12 +32,11 @@ enum struct ConvOutputLayout
     NHWK, // 1
 };
 
-int profile_conv_fwd_bias_relu_add(int argc, char* argv[])
+int profile_conv_fwd_bias_relu(int argc, char* argv[])
 {
     if(argc != 25)
     {
-        printf(
-            "arg1: tensor operation (conv_fwd_bias_relu_add: ForwardConvolution+Bias+ReLu+Add)\n");
+        printf("arg1: tensor operation (conv_fwd_bias_relu: ForwardConvolution+Bias+ReLu)\n");
         printf("arg2: data type (0: fp32; 1: fp16)\n");
         printf("arg3: input tensor layout (0: NCHW; 1: NHWC)\n");
         printf("arg4: weight tensor layout (0: KCYX; 1: KYXC)\n");
@@ -86,13 +85,13 @@ int profile_conv_fwd_bias_relu_add(int argc, char* argv[])
     if(data_type == ConvDataType::F16_F16_F16 && in_layout == ConvInputLayout::NHWC &&
        wei_layout == ConvWeightLayout::KYXC && out_layout == ConvOutputLayout::NHWK)
     {
-        ck::profiler::profile_conv_fwd_bias_relu_add_impl<2,
-                                                          ck::half_t,
-                                                          ck::half_t,
-                                                          ck::half_t,
-                                                          ck::tensor_layout::convolution::NHWC,
-                                                          ck::tensor_layout::convolution::KYXC,
-                                                          ck::tensor_layout::convolution::NHWK>(
+        ck::profiler::profile_conv_fwd_bias_relu_impl<2,
+                                                      ck::half_t,
+                                                      ck::half_t,
+                                                      ck::half_t,
+                                                      ck::tensor_layout::convolution::NHWC,
+                                                      ck::tensor_layout::convolution::KYXC,
+                                                      ck::tensor_layout::convolution::NHWK>(
             do_verification,
             init_method,
             do_log,
