+++ library/src/profiler/profile_gemm.cpp
@@ -6,7 +6,7 @@
 #include <initializer_list>
 #include <cstdlib>
 
-#include "profiler/include/profile_gemm_impl.hpp"
+#include "profile_gemm_impl.hpp"
 
 enum struct GemmMatrixLayout
 {
