+++ library/src/profiler/profile_gemm_add_add_fastgelu.cpp
@@ -6,7 +6,7 @@
 #include <initializer_list>
 #include <cstdlib>
 
-#include "profiler/include/profile_gemm_add_add_fastgelu_impl.hpp"
+#include "profile_gemm_add_add_fastgelu_impl.hpp"
 
 int profile_gemm_add_add_fastgelu(int argc, char* argv[])
 {
