+++ library/src/profiler/profile_gemm_bias_relu.cpp
@@ -6,7 +6,7 @@
 #include <initializer_list>
 #include <cstdlib>
 
-#include "profiler/include/profile_gemm_bias_relu_add_impl.hpp"
+#include "profile_gemm_bias_relu_impl.hpp"
 
 enum struct GemmMatrixLayout
 {
@@ -26,11 +26,11 @@ enum struct GemmDataType
     F16_F16_F16, // 1
 };
 
-int profile_gemm_bias_relu_add(int argc, char* argv[])
+int profile_gemm_bias_relu(int argc, char* argv[])
 {
-    if(!(argc == 15 || argc == 16))
+    if(!(argc == 14 || argc == 15))
     {
-        printf("arg1: tensor operation (gemm: GEMM+Bias+ReLU+Add)\n");
+        printf("arg1: tensor operation (gemm: GEMM+Bias+ReLU)\n");
         printf("arg2: data type (0: fp32; 1: fp16)\n");
         printf("arg3: matrix layout (0: A[m, k] * B[k, n] = C[m, n];\n");
         printf("                     1: A[m, k] * B[n, k] = C[m, n];\n");
@@ -40,8 +40,8 @@ int profile_gemm_bias_relu_add(int argc, char* argv[])
         printf("arg5: initialization (0: no init; 1: integer value; 2: decimal value)\n");
         printf("arg6: print tensor value (0: no; 1: yes)\n");
         printf("arg7: time kernel (0=n0, 1=yes)\n");
-        printf("arg8 to 14: M, N, K, StrideA, StrideB, StrideC, StrideC1\n");
-        printf("arg15: split k into  mulitiple batch\n");
+        printf("arg8 to 13: M, N, K, StrideA, StrideB, StrideC\n");
+        printf("arg14: split k into  mulitiple batch\n");
         exit(1);
     }
 
@@ -56,19 +56,18 @@ int profile_gemm_bias_relu_add(int argc, char* argv[])
     const int N = std::stoi(argv[9]);
     const int K = std::stoi(argv[10]);
 
-    const int StrideA  = std::stoi(argv[11]);
-    const int StrideB  = std::stoi(argv[12]);
-    const int StrideC  = std::stoi(argv[13]);
-    const int StrideC1 = std::stoi(argv[14]);
+    const int StrideA = std::stoi(argv[11]);
+    const int StrideB = std::stoi(argv[12]);
+    const int StrideC = std::stoi(argv[13]);
 
     if(data_type == GemmDataType::F16_F16_F16 && layout == GemmMatrixLayout::MK_KN_MN)
     {
-        ck::profiler::profile_gemm_bias_relu_add_impl<ck::half_t,
-                                                      ck::half_t,
-                                                      ck::half_t,
-                                                      ck::tensor_layout::gemm::RowMajor,
-                                                      ck::tensor_layout::gemm::RowMajor,
-                                                      ck::tensor_layout::gemm::RowMajor>(
+        ck::profiler::profile_gemm_bias_relu_impl<ck::half_t,
+                                                  ck::half_t,
+                                                  ck::half_t,
+                                                  ck::tensor_layout::gemm::RowMajor,
+                                                  ck::tensor_layout::gemm::RowMajor,
+                                                  ck::tensor_layout::gemm::RowMajor>(
             do_verification,
             init_method,
             do_log,
@@ -78,17 +77,16 @@ int profile_gemm_bias_relu_add(int argc, char* argv[])
             K,
             (StrideA < 0) ? K : StrideA,
             (StrideB < 0) ? N : StrideB,
-            (StrideC < 0) ? N : StrideC,
-            (StrideC1 < 0) ? N : StrideC1);
+            (StrideC < 0) ? N : StrideC);
     }
     else if(data_type == GemmDataType::F16_F16_F16 && layout == GemmMatrixLayout::MK_NK_MN)
     {
-        ck::profiler::profile_gemm_bias_relu_add_impl<ck::half_t,
-                                                      ck::half_t,
-                                                      ck::half_t,
-                                                      ck::tensor_layout::gemm::RowMajor,
-                                                      ck::tensor_layout::gemm::ColumnMajor,
-                                                      ck::tensor_layout::gemm::RowMajor>(
+        ck::profiler::profile_gemm_bias_relu_impl<ck::half_t,
+                                                  ck::half_t,
+                                                  ck::half_t,
+                                                  ck::tensor_layout::gemm::RowMajor,
+                                                  ck::tensor_layout::gemm::ColumnMajor,
+                                                  ck::tensor_layout::gemm::RowMajor>(
             do_verification,
             init_method,
             do_log,
@@ -98,17 +96,16 @@ int profile_gemm_bias_relu_add(int argc, char* argv[])
             K,
             (StrideA < 0) ? K : StrideA,
             (StrideB < 0) ? K : StrideB,
-            (StrideC < 0) ? N : StrideC,
-            (StrideC1 < 0) ? N : StrideC1);
+            (StrideC < 0) ? N : StrideC);
     }
     else if(data_type == GemmDataType::F16_F16_F16 && layout == GemmMatrixLayout::KM_KN_MN)
     {
-        ck::profiler::profile_gemm_bias_relu_add_impl<ck::half_t,
-                                                      ck::half_t,
-                                                      ck::half_t,
-                                                      ck::tensor_layout::gemm::ColumnMajor,
-                                                      ck::tensor_layout::gemm::RowMajor,
-                                                      ck::tensor_layout::gemm::RowMajor>(
+        ck::profiler::profile_gemm_bias_relu_impl<ck::half_t,
+                                                  ck::half_t,
+                                                  ck::half_t,
+                                                  ck::tensor_layout::gemm::ColumnMajor,
+                                                  ck::tensor_layout::gemm::RowMajor,
+                                                  ck::tensor_layout::gemm::RowMajor>(
             do_verification,
             init_method,
             do_log,
@@ -118,17 +115,16 @@ int profile_gemm_bias_relu_add(int argc, char* argv[])
             K,
             (StrideA < 0) ? M : StrideA,
             (StrideB < 0) ? N : StrideB,
-            (StrideC < 0) ? N : StrideC,
-            (StrideC1 < 0) ? N : StrideC1);
+            (StrideC < 0) ? N : StrideC);
     }
     else if(data_type == GemmDataType::F16_F16_F16 && layout == GemmMatrixLayout::KM_NK_MN)
     {
-        ck::profiler::profile_gemm_bias_relu_add_impl<ck::half_t,
-                                                      ck::half_t,
-                                                      ck::half_t,
-                                                      ck::tensor_layout::gemm::ColumnMajor,
-                                                      ck::tensor_layout::gemm::ColumnMajor,
-                                                      ck::tensor_layout::gemm::RowMajor>(
+        ck::profiler::profile_gemm_bias_relu_impl<ck::half_t,
+                                                  ck::half_t,
+                                                  ck::half_t,
+                                                  ck::tensor_layout::gemm::ColumnMajor,
+                                                  ck::tensor_layout::gemm::ColumnMajor,
+                                                  ck::tensor_layout::gemm::RowMajor>(
             do_verification,
             init_method,
             do_log,
@@ -138,8 +134,7 @@ int profile_gemm_bias_relu_add(int argc, char* argv[])
             K,
             (StrideA < 0) ? M : StrideA,
             (StrideB < 0) ? K : StrideB,
-            (StrideC < 0) ? N : StrideC,
-            (StrideC1 < 0) ? N : StrideC1);
+            (StrideC < 0) ? N : StrideC);
     }
     else
     {
