+++ library/src/profiler/profile_reduce.cpp
@@ -13,8 +13,8 @@
 
 #include "ck/library/host_tensor/host_common_util.hpp"
 
-#include "profiler/include/profile_reduce_impl.hpp"
-#include "profiler/include/data_type_enum.hpp"
+#include "profile_reduce_impl.hpp"
+#include "data_type_enum.hpp"
 
 using namespace std;
