--- profiler/include/profile_convnd_fwd.hpp
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: MIT
-// Copyright (c) 2018-2022, Advanced Micro Devices, Inc. All rights reserved.
-
-#pragma once
-
-namespace ck {
-namespace profiler {
-
-int profile_convnd_fwd(int argc, char* argv[]);
-
-} // namespace profiler
-} // namespace ck
